backend/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
psycopg2-binary==2.9.9
gunicorn==22.0.0
dj-database-url==2.1.0
django-cors-headers==4.3.1
